Bug fixes and code changes

+++ Experiments/msudipto/UNIT-2/springboot_unit2_1_onetoone/springboot_example/src/main/java/onetoone/Persons/Person.java
@@ -56,7 +56,7 @@ public class Person {
 
     // =============================== Getters and Setters for each field ================================== //
 
-    public long getId() {  // Fixing return type to int
+    public int getId() {
         return id;
     }
